Fix issue #62.
https://github.com/Andersbakken/rtags/issues/62

We don't get references to templatized superclass members in template member functions
E.g.

template
class A
{
public:
int size() const { return 1; }
};

template
class B : public A
{
public:
bool isEmpty() const { return A::size(); }
};

jumping from the call to size() won't go anywhere.

This is because there simply isn't a cursor pointing to it available to us.

This fix hacks around this by looking up the name of the symbol we're
on, getting the reference to the class template and finding a child
with same name. If there's overloads or something to that effect we
might plain be wrong.