Hopefully provide enough infrastructure to take care of issue #92.
It's now possible to override project for rc --compile commands with
--with-project.

It should now be possible to do the following:

RTAGS_PROJECT=/path/to/project make

and all source files compiled should be placed under the
/path/to/project project.

Similarly in elisp one could do something like this:

(defun foobar (arg)
  "/path/to/project")

(setq rtags-match-source-file-to-project 'foobar)

And all requests should go into that project.

(The elisp part can kinda be skipped, it's mostly there to make sure
projects automatically gets loaded. If the project already is loaded
we should be able to determine which one it is based on the file alone.