FUNCTIONS TO START/STOP THE RDM PROCESS FROM EMACS.
To start the process it uses the function `rtags-includes-func` for settings the includes to the `rdm` process. The default function just return the variable `rtags-rdm-includes`. The following function are defined:

`rtags-start-process-maybe` : start the rdm process if not started.
`rtags-cancel-process` : stop the rdm process.
`rtags-restart-process` : stop and start the rdp process.