minor fix for rtags-get-summary-text
rtags-display-summary sometimes gets errors when the "cursorinfo" is
nil(e.g., when the point is inside "#include <iostream>"); the change is
simply adding a check inside rtags-get-summary-text. However, the better
way for this case is to display the location of file `iostream'.

Another change is to replace the keymap of rtags-print-cursorinfo for
rtags-display-summary.