Use readlink instead of a Python one-liner
Use readlink instead of python -c "import os, sys; print(os.path.realpath(\"$1\"))".
This removes an unnecessary dependency upon Python as well as an
injection-attack vector: followSymlink can only be used for data that is
trusted absolutely, because a value of '"; ... "' for $1 (the argument
to followSymlink), where '...' is arbitrary Python code, will cause the
input to be executed.