Long live 2.0.
We've changed the architecture to store almost all data in files that
are mmap'ed in. There's a slight performance degradation for queries but
a number of other benefits:

Vastly reduced memory usage.
Slightly increased performance for indexing.
Much better functionality during indexing. Since we avoid the sync step
data should now be usable before everything's indexed. This also means
that RTags can determine an appropriate order of indexing for files
based on which files are open in your editor.