Adding rtags-get-include-file-for-symbol
rtags-get-include-file-for-symbol uses a mechanism very similar to rtags-find-symbol-internal to list all definitions of a symbol. The list variable rtags-include-prefixes is used to shorten the include paths. For example, if "abc" is an element of rtags-include-prefixes and the symbol is defined in "path/to/file/in/abc/rest/of/path", rtags-get-include-file-for-symbol will insert "#include abc/rest/of/path" into the current buffer.