Support elisp installation from out-of-source builds.
When using an out-of-source build style, the generated installer would
look in the source for elisp (and compiled elisp) files rather than in
the build directory.

This eliminates the "configure_file" step that had no effect anyway and
replaces it with an install directive at that site.  The compiled elisp
custom command also learns how to do an install.