Use pkg-config to figure out if bash completion support is available
- If PKG was not found we fail silently
- If PKG could not find bash-completion support, we fail silently, since
  bash completion is not a must
- Use `/usr` as default install path if the user has not provided a path
  as argument, this is the default install location.

This is the best option to install the bash completion files, right now.
We should mention in the README, if the user wants bash completion
support he/she needs to make sure that the `pkg-config` tool and the
`bash-completion` package is installed. Also this makes it easier, in
case rtags gets packaged as `rpm` or `deb` file one they or another.