Add possibility to force bash completion installation
The bash completion files can now be forced to be included in the
installation process, by setting the variable
`FORCE_BASH_COMPLETION_INSTALLATION` to `TRUE` or `ON`.
By default this variable is set to `FALSE` (`OFF`). If it set to TRUE
ignore pkg-config at all. The variable can also be set through the
interactive `ccmake` environment.

Example:

Default value of `BASH_COMPLETION_COMPLETIONSDIR` is
"/usr/share/bash-completion/completions" if installation is forced, but
can of course be overridden.

```bash
cmake -DFORCE_BASH_COMPLETION_INSTALLATION=ON
```

The value of `CMAKE_INSTALL_PREFIX` will always be the prefix of
`BASH_COMPLETION_COMPLETIONSDIR`.