Quick and dirty solution for escaping briefComment.
M-x rtags-find-symbol used to fail with: Invalid read syntax: "#",
when the comment of the function was crafted in a particular way.
The problem was that briefComment of the function ended with a backslash.
This would cause the rdm to genaret elisp: (cons 'briefComment "...\"),
which accidentally escapes the right double-quotes, leading to bad lisp.
A proper solution would be to escape the briefComment string.
The current patch adds a single whitespace at the end of the briefComment
string, making sure that the right double-quotes do not get escaped.

An example code that reproduces the error:
(If we place the cursor on var and try to rtags-find-symbol)

/*********\
 Comments
\*********/
int foo() {
  int var;
}

Before the patch rdm used to generate this:
(cons 'briefComment "*********\")
After the patch it generates this:
(cons 'briefComment "*********\ ")