Support XDG basedir specification
Construct the default config and data directories, as well as the
location of the socket file per XDG basedir specification.

Cache and data dirs are basically the same: if the previously defined
default directory exists it is used. This way there is no change with
existing setups. Otherwise the path is constructed from the XDG
environment variable or the default if that doesn't exist:

$XDG_CONFIG_DIR/rtags/<app>rc  $XDG_CONFIG_DIR default is ~/.config
$XDG_CACHE_HOME/rtags          $XDG_CACHE_HOME default is ~/.cache

The socket file is slightly different, since XDG does not specify a
default runtime directory. If $XDG_RUNTIME_DIR does not exist, then
the current default is used. Otherwise, the socket file is created in
$XDG_RUNTIME_DIR