emacs: Implement 'rename with multiple cursors'
This commit adds the ability to rename symbol at point
using multiple-cursors in emacs, a feature from IDEs.
When the scope of the symbol is just one file, then
multiple-cursors can be used to edit multiple occurances
of this symbol.
New customization option was created to enable or disable
that feature (rtags-use-multiple-cursors).
By default this feature is disabled.