Use equal to compare return code of rc calls
The return value of process calls can by any lisp object. Hence, when
comparing the return value, which is a lisp object, using `=`, `=` will fail
when the returned lisp object is something other then a number or marker,
like a string lisp object. This can be the case when the called process exited
abnormal, as it happened in #1187.

If the return value is "Aborted" print an error message that the program rc
exited abnormal.