Use --strip-paren to exclude function variables from --find-symbols
When --find-symbols is used with --wildcard-symbol-names, the results
may include function variables (which do not match the search pattern).
Using --match-regexp with a lookahead to exclude the ')::' function
variable signature is very slow.

When --strip-paren is used with --list-symbols, it excludes function
variables as a side effect.
This change aligns --find-symbols with that behavior.