Consider all other unsaved files when code completing a given file
If the file being code-completed depends on other unsaved header files,
then those other headers also need to be reindexed. Otherwise, when
the file being code-completed is reindexed, the last saved versions of
those header dependencies will be used.

This change is a work in progress. It introduces some inefficiencies
that should be optimized:
- It passes all unsaved files to the completion thread, rather than only
those that the file being code-completed depends on.
- It copies and compares UnsavedFiles (std::unordered_map) objects
several times.