[travis] Find llvm-config using find
In homebrew llvm has changed and that is why FindLibClang.cmake cannot find
llvm-config.
  llvm-7.0.0.high_sierra.bottle.1.tar.gz --> llvm-7.0.0_1.high_sierra.bottle.tar.gz

I'm not quite sure right now if that is just in travis that case or in
general. If in general than we would need to adopt FindLibClang.cmake

Nose version also contains now "_1" at the end of the version string