Consider unsaved files when reindexing to recover from #include errors
If unsaved files are not passed to an indexing job for those same files,
then the last saved versions of the files are used. This change fixes the
following issue caused by that behavior:

1. An #include error is introduced into a file, and the file is saved
and reindexed.
2. The #include error is removed, and the file is reindexed again with
--reindex --unsaved-file, but not saved again.
3. After the indexing job has finished, updating dependencies causes the
file to be reindexed yet again (because the #include error has been
removed).
4. The last saved version of the file (still containing the #include
error) is used when updating dependencies.