Ensure that the project path always has a trailing slash
This change fixes restoring projects when rdm is stopped and restarted
with the --no-realpath option. It also causes project paths to be
displayed consistently whether or not --no-realpath is specified.

The underlying issue is that Path::resolve() adds a trailing slash when
realpath is enabled, but does not when realpath is disabled.