Only push the tag not the changes when bumping version
When the tag build fails, we have already pushed the changes to master. This is
bad because no releases will be created. Instead only push the tag to the
repo, and later when the build succeeds push to the master branch. We don't
care if the build is than failing on the master branch. Since the release have
already been created and is available. So the MELPA package can be safely
updated.