Add when argument to define-obsolete-function-alias
The `when` has become mandatory since
https://emba.gnu.org/emacs/emacs/-/commit/32c6732d16385f242b1109517f25e9aefd6caa5c (credit
to @zflat).

Fortunately the emacs-24.3 (the lowest supported version according to package),
has already support for the `when` parameter.

I've used `v2.2` as this is the tag that follows commit that added the obsolete
call (dbf1ca3b2518d43f8c1b897d71126cfd132ef30b).